apikeyclient/pyproject.toml
[build-system]
requires = ["setuptools", "wheel"]

[project]
name = "datadiensten-apikeyclient"
version = "0.6.0"
description = "Client for Datadiensten apikeyserv"
dependencies = ['pause>=0.3', 'requests', 'PyJWT[crypto]==2.8.0', 'build']

[project.optional-dependencies]  
dev = [
    'build', 
    'twine'
]
test = ['requests-mock']

[project.urls]
"Homepage" = "https://github.com/Amsterdam/apikeyserv"

[tool.isort]
profile = "black"
line_length = 99

[tool.black]
line-length = 99
exclude = '''
/(
    \.git
  | \.mypy_cache
  | \.tox
  | \.venv
  | _build
  | build
  | dist
)/
'''

github_url = "https://github.com/Amsterdam/apikeyserv"
